video: tegra: camera: change sclk init value
Change initial sclk value from 204MHz to 80MHz.
It helps reduce power while maintaining perf.

Bug 1234580

Reviewed-on: http://git-master/r/205973
GVS: Gerrit_Virtual_Submit

+++ drivers/video/tegra/camera/camera_priv_defs.h
@@ -93,7 +93,7 @@ struct clock_data {
 	int index;
 	char *name;
 	bool init;
-	int freq;
+	unsigned long freq;
 };
 
 struct tegra_camera *tegra_camera_register(struct platform_device *ndev);
